+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CHANGELOG for Sulu
 
 * dev-develop
     * FEATURE     #1584 [ContentBundle]   Show UI elements in content management based on security
+    * BUGFIX      #1633 [Rest]            Listbuilder: fix for concat joins in where clause;
+                                          Fix in sort (order by id as default)
     * BUGFIX      #1625 [MediaBundle]     Fixed queries for PostgreSQL
     * ENHANCEMENT #1626 [AdminBundle]     Added require-css to load css dynamically with require
     * BUGFIX      #1619 [CategoryBundle]  Fixed category view cropping responsive